% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gf_AgregArbres}
\alias{gf_AgregArbres}
\title{Aggrégation des arbres à l'échelle de la placette.}
\usage{
gf_AgregArbres()
}
\value{
La fonction construit les tables suivantes. Elles sont enregistrées dans le dossier Tables

gfDen, gfPla, gfPlaDen, gfPlaCat, gfPlaCatDen, gfPlaCatEssReg, gfPlaCatEssRegDen,
gfPlaEssReg, gfPlaEssRegDen, gfPlaEssRegQual, gfPlaEssRegQualDen, gfPlaCatQual1,
gfPlaCatQual2, gfPlaCatQual2Den, gfPlaEss, gfPlaEssDen, gfPlaClasseQual1,
gfPlaClasseQual1Den, gfPlaClasseQual2, gfPlaClasseQual2Den, gfPlaClasseEssReg,
gfPlaEssRegQual1, gfPlaEssRegQual2, gfPerchesPla, gfPerchesPlaEss, gfPerchesPlaQual,
gfTaillisPla, gfTaillisPlaEss, gfTaillisPlaEssReg, gfRegePla, gfRegePlaEss, gfRegePlaEssReg
}
\description{
Sous-programme permettant d'aggréger toutes les variables à l'échelle de la placette.
Il fournit de nombreux tableaux croisés.
}
\author{
Bruciamacchie Max
}

