% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gf_AgregPlacettes}
\alias{gf_AgregPlacettes}
\title{Aggrégation des placettes à une échelle choisie par l'opérateur.}
\usage{
gf_AgregPlacettes()
}
\value{
La fonction construit les tables suivantes. Elles sont enregistrées dans le dossier Tables

gffNGVdisp, gffNGVstrate, gffNGVcc,
gffNGVdispEss, gffNGVdispEssReg, gffNGVdispCat,
gffHistRB, gffHistStrate, gffHistRBEssReg,
gffBMRB, gffBMRBstadeD, gffBMRBstadeE
}
\description{
Sous-programme permettant de calculer les moyennes, coefficient de variations et erreurs relatives par groupes de placettes
tels que choisis dans la feuille Placettes.
}
\author{
Bruciamacchie Max
}

