% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gf_Calculs}
\alias{gf_Calculs}
\title{Calcul des variables dendrométriques, économiques et écologiques.}
\usage{
gf_Calculs(TauxR = 0.03)
}
\arguments{
\item{TauxR}{= taux d'actualisation (0.03 par défaut)}
}
\value{
La fonction construit les tables suivantes.

arbres, Reges, Taillis, Reperes, BMSLineaires, BMSsup30, BMP, Codes

Elles sont enregistrées dans le dossier Tables sous le nom : gfTablesBrutes.RData.
}
\description{
Cette fonction commence par calculer le poids de chaque arbre,
puis pour chacun d'eux, calcule les variables dendrométriques, économiques et écologiques ramenées à l'hectare.
}
\author{
Bruciamacchie Max
}

