% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gf_Tables2Xls}
\alias{gf_Tables2Xls}
\title{Export des tables elaborees au format xlsx.}
\usage{
gf_Tables2Xls()
}
\value{
La fonction exporte dans un classeur au format .xlsx les tables suivantes :

gfDendroForet,gfDendroGroupe,gfDendroGroupe1,gfDendroGroupe2,gfDendroStation,
gfDendroTypo,gfDendroStrate,gfEcoForet,gfEcoGroupe,gfEcoGroupe1,gfEcoGroupe2,
gfEcoStation,gfEcoTypo,gfEcoStrate,gfEssMoy,gfEssRegMoy,gfTaillisEss,gfTaillisEssReg,
gfRegeEss,gfHistClasseEssReg,gfHistClasse,gfHistClasseABC,gfHistClasseD,gfCatMoy

Le classeur est enregistré dans le dossier Out/Resultats sous le nom : TabElaborees.xlsx.
}
\description{
Sauvegarde des tables élaborées dans un classeur Excel au format xlsx, dans un dossier Out.
}
\author{
Bruciamacchie Max
}

