% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gf_Xls2Rdata}
\alias{gf_Xls2Rdata}
\title{Conversion des données brutes de la base GF en 2 archives RData.}
\usage{
gf_Xls2Rdata(file)
}
\arguments{
\item{rep}{= répertoire contenant les données}
}
\value{
La fonction exporte dans le dossier Tables l'archive suivante :

gfDonneesBrutes.Rdata
}
\description{
Import des classeurs Excel de chaque forêt et sauvegarde dans l'archive
gfDonneesBrutes.Rdata.

Import des différentes tables de codification et sauvegarde dans l'archive psdrfCodes.Rdata.
}
\author{
Bruciamacchie Max
}

